Create 2024-11-03-Research-Or-Build
First attempt at answering to research or build in your youth. There's some goals of the paper and some other thoughts. Sectioned off at a good written portion, essay goals, and badly written portion. But the good part may not be relevant as the flavor of the essay is changing..